Standardoze Type.platformSpecific API for type calculation between Scala 2 and 3 (not by sharing interface, but by sharing conventions)
